move state to lib
