+++ src/App.svelte
@@ -1,7 +1,7 @@
 <script>
     import { onMount } from "svelte";
     import { navigate, Router } from "svelte-routing";
-    import { userState } from "./shared/state/userState.svelte.js";
+    import { userState } from "./lib/state/userState.svelte.js";
     import Header from "./Header.svelte";
     import Routes from "./routes/Routes.svelte";
     import "./app.css";
