+++ src/Header.svelte
@@ -1,7 +1,7 @@
 <script>
     import { Link } from "svelte-routing";
     import routes from "./routes/routes.js";
-    import { userState } from "./shared/state/userState.svelte.js";
+    import { userState } from "./lib/state/userState.svelte.js";
 
     let isMenuActive = $state(false);
     let userDropdownActive = $state(false);
