+++ src/routes/Logout.svelte
@@ -1,7 +1,7 @@
 <script>
     import { onMount } from "svelte";
     import { navigate } from "svelte-routing";
-    import { userState } from "../shared/state/userState.svelte.js";
+    import { userState } from "../lib/state/userState.svelte.js";
     import { logout } from "../lib/controllers/authentication.js";
     import routes from "./routes.js";
